+++ packages/core/src/models/project.ts
@@ -321,6 +321,10 @@ export class ProjectModel {
     return finder(this.pages);
   }
 
+  getHomepage() {
+    return this.homepage ? this.getPage(this.homepage) : this.getPages()[0];
+  }
+
   /**
    * 新建页面
    * @param page
@@ -599,6 +603,7 @@ export class ProjectModel {
     const name = upperFirstCamelCase(block.name);
     block.id = id;
     block.type = 'block';
+    block.fromType = block.fromType || 'Schema';
     block.dsl = new BlockModel({ id, name }).toDsl();
     this.blocks.push(block);
     const type = block.fromType || 'Schema';