+++ packages/designer/src/framework/designer.ts
@@ -55,6 +55,7 @@ export interface DesignHelper {
   rect: DOMRect;
   type?: DropPosition;
   path?: Array<NodeModel | BlockModel>;
+  indexes?: number[];
 }
 
 export class Designer {
@@ -345,7 +346,7 @@ export class Designer {
   }
 
   private getNodeByElement(el: VtjElement): NodeModel | null {
-    const id = el.__vtj__ ?? '';
+    const id = (el as HTMLElement).getAttribute('data-vtj') || el.__vtj__ || '';
     return NodeModel.nodes[id] || null;
   }
 
@@ -377,6 +378,16 @@ export class Designer {
     return [...nodePath, root];
   }
 
+  private getNodePathIndex(nodePath: Array<BlockModel | NodeModel>) {
+    return nodePath.map((n) => {
+      if (isBlock(n)) {
+        return 0;
+      } else {
+        return n.parent?.findChildIndex(n) || 0;
+      }
+    });
+  }
+
   private setDslFrom(dsl: NodeSchema) {
     const desc = this.engine.assets.componentMap.get(dsl.name);
     dsl.from = dsl.from || desc?.package;
@@ -406,7 +417,9 @@ export class Designer {
       this.document.querySelectorAll('*')
     );
     return list.find(
-      (el) => (el as VtjElement).__vtj__ === model.id
+      (el) =>
+        (el as VtjElement).__vtj__ === model.id ||
+        (el as any).getAttribute(`data-vtj`) === model.id
     ) as VtjElement;
   }
 
@@ -433,12 +446,14 @@ export class Designer {
     const rect = el.getBoundingClientRect();
     const type = this.getDropType(rect, e.clientX, e.clientY);
     const path = this.getNodePath(targets);
+    const indexes = this.getNodePathIndex(path);
     return {
       el,
       model,
       rect,
       type,
-      path
+      path,
+      indexes
     };
   }