+++ dev/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.16.20](https://github.com/samchen08/vtj.pro/compare/dev-web@0.16.19...dev-web@0.16.20) (2026-04-18)
+
+**Note:** Version bump only for package dev-web
+
+
+
+
+
+## [0.16.19](https://github.com/samchen08/vtj.pro/compare/dev-web@0.16.18...dev-web@0.16.19) (2026-04-16)
+
+**Note:** Version bump only for package dev-web
+
+
+
+
+
+## [0.16.18](https://github.com/samchen08/vtj.pro/compare/dev-web@0.16.17...dev-web@0.16.18) (2026-04-15)
+
+**Note:** Version bump only for package dev-web
+
+
+
+
+
+## [0.16.17](https://github.com/samchen08/vtj.pro/compare/dev-web@0.16.16...dev-web@0.16.17) (2026-04-13)
+
+**Note:** Version bump only for package dev-web
+
+
+
+
+
+## [0.16.16](https://github.com/samchen08/vtj.pro/compare/dev-web@0.16.15...dev-web@0.16.16) (2026-04-10)
+
+**Note:** Version bump only for package dev-web
+
+
+
+
+
 ## [0.16.15](https://github.com/samchen08/vtj.pro/compare/dev-web@0.16.14...dev-web@0.16.15) (2026-04-07)
 
 **Note:** Version bump only for package dev-web